+++ tests/frameservers/clockreq/clockreq.c
@@ -4,6 +4,7 @@
 #include <unistd.h>
 #include <stdbool.h>
 #include <string.h>
+#include <inttypes.h>
 
 #include <arcan_shmif.h>
 
@@ -12,22 +13,39 @@ int main(int argc, char** argv)
 	struct arg_arr* aarr;
 	struct arcan_shmif_cont cont = arcan_shmif_open(
 		SEGID_APPLICATION, SHMIF_ACQUIRE_FATALFAIL, &aarr);
+	bool resubmit = false;
 
 	arcan_event ev = {
-		.ext.kind = ARCAN_EVENT(CLOCKREQ),
-		.ext.clock.rate = 2,
-		.ext.clock.dynamic = (argc > 1 && strcmp(argv[1], "dynamic") == 0)
-	};
-	arcan_shmif_enqueue(&cont, &ev);
+			.ext.kind = ARCAN_EVENT(CLOCKREQ),
+			.ext.clock.rate = 50,
+			.ext.clock.id = 10,
+		};
 
-	ev.ext.clock.dynamic = false;
-	int tbl[] = {20, 40, 42, 44, 60, 80, 86, 88, 100, 120};
-	int step = 0;
-
-	for (size_t i=0; i < sizeof(tbl)/sizeof(tbl[0]); i++){
-		ev.ext.clock.once = true;
-		ev.ext.clock.rate = tbl[i];
-		ev.ext.clock.id = i + 2; /* 0 index and 1 is reserved */
+	if (argc > 1 && strcmp(argv[1], "present") == 0){
+		arcan_shmif_enqueue(&cont, &(struct arcan_event){
+				.ext.kind = ARCAN_EVENT(CLOCKREQ),
+				.ext.clock.dynamic = 1
+		});
+		printf("requesting feedback on submit-ack and present\n");
+		arcan_shmif_signal(&cont, SHMIF_SIGVID);
+		resubmit = true;
+	}
+	else if (argc > 1 && strcmp(argv[1], "vsignal") == 0){
+		cont.hints = SHMIF_RHINT_VSIGNAL_EV;
+		arcan_shmif_resize(&cont, cont.w, cont.h);
+		arcan_shmif_signal(&cont, SHMIF_SIGVID);
+		resubmit = true;
+	}
+	else if (argc > 1 && strcmp(argv[1], "vblank") == 0){
+		arcan_shmif_enqueue(&cont, &(struct arcan_event){
+				.ext.kind = ARCAN_EVENT(CLOCKREQ),
+				.ext.clock.dynamic = 2
+		});
+		printf("requesting timer each blank\n");
+		arcan_shmif_signal(&cont, SHMIF_SIGVID);
+	}
+	else {
+		printf("running basic [n=50@25Hz -> 2s] custom timer with id 10\n");
 		arcan_shmif_enqueue(&cont, &ev);
 	}
 
@@ -35,16 +53,10 @@ int main(int argc, char** argv)
 		if (ev.category == EVENT_TARGET)
 			switch(ev.tgt.kind){
 				case TARGET_COMMAND_STEPFRAME:
-					printf("step: %d, source: %d\n",
-						ev.tgt.ioevs[0].iv, ev.tgt.ioevs[1].iv);
-					if (ev.tgt.ioevs[1].iv > 1){
-						if (step == ev.tgt.ioevs[1].iv-2)
-							printf("custom timer %d OK\n", step);
-						else
-							printf("timer out of synch, expected %d got %d\n",
-							step, ev.tgt.ioevs[1].iv-2);
-						step++;
-					}
+					printf("step: %d, source: %d cval: %"PRIu32"\n",
+						ev.tgt.ioevs[0].iv, ev.tgt.ioevs[1].iv, ev.tgt.ioevs[2].uiv);
+					if (resubmit)
+						arcan_shmif_signal(&cont, SHMIF_SIGVID);
 				break;
 				case TARGET_COMMAND_EXIT:
 					goto done; /* break(1), please */